Item management changes - CHANGELOG not updated
git-svn-id: http://www.linux-addicted.net/repos/tinns@90 ef8e7eb3-8d0b-0410-9eab-9562649fb893

+++ trunk/tinns-v2/src/game/include/msgdecoder.h
@@ -54,7 +54,7 @@ struct PMsgDecodeData
 	  u8 mState;
 	  u8 mUnknownType;
 	  bool mHandling0x13Sub;
-	  u8 Sub0x13Start;
+	  u16 Sub0x13Start;
 	  u16 Sub0x13StartNext;
 	  std::stringstream mName;
 	  std::string mErrorDetail;
